+++ docs/source/changelog.md
@@ -8,6 +8,8 @@ Release notes for the `space_packet_parser` library
 ### v5.1.0 (unreleased)
 - BUGFIX: Fix kbps calculation in packet generator for showing progress.
 - Add support for string and float encoded enumerated lookup parameters.
+- Add an Excel spreadsheet definition to XTCE conversion utility to enable a quicker way
+  to get started with XTCE definitions. The current format supports GSEOS-style packet definitions.
 
 ### v5.0.1 (released)
 - BUGFIX: Allow raw_value representation for enums with falsy raw values. Previously these defaulted to the enum label.